Upstreaming changes in Netvirt
(Dependent on https://git.opendaylight.org/gerrit/#/c/42540/)

This commit includes VPN Engine Concurrency Fixes also some SNAT and DNAT
regression issues.

1. We have made multiple fixes here for safe concurrency.

2. We have moved all helper methods that deal with VRFEntry
resource into FibManager from VPNManager.

3. We have gracefully handled deletion of bgpvpn regardless of
how many ever external routes are available in such vpn.

4. We have fixed NATService to directly use FibManager to
program SNAT/DNAT VRFEntries.

5. As of this review, all local VM routes WILL NOT be batched for
write into VRFEntry Config DataStore.

6. All the external routes imported from BGP, WILL BE batched and
written to VRFEntry Config DataStore.

+++ vpnservice/natservice/natservice-impl/src/main/java/org/opendaylight/netvirt/natservice/internal/FloatingIPListener.java
@@ -665,7 +665,7 @@ public class FloatingIPListener extends AbstractDataChangeListener<IpMapping> im
         return null;
     }
 
-    void updateOperationalDS(String routerId, String interfaceName, int label, String internalIp, String externalIp) {
+    void updateOperationalDS(String routerId, String interfaceName, long label, String internalIp, String externalIp) {
 
         LOG.info("Updating operational DS for floating ip config : {} with label {}", internalIp, label);
         InstanceIdentifier<Ports> portsId = NatUtil.getPortsIdentifier(routerId, interfaceName);
